Move issues agent into contrib/maintenance shared package
Split contrib/issues_agent/tasks.py into a package of reusable modules —
datasets.py (repo path resolution), archive.py (data.opensanctions.org
versions and artifacts), issues.py (normalization/checksum/grouping),
github.py (outage board, autofix PR checks) — with issues_agent.py keeping
only agent policy: model/turn selection, branch naming, PR dedup and matrix
emission. This makes the mechanics shareable with other maintenance tools,
starting with the diagnose report added in the next commit.

The workflow now invokes it as `python -m contrib.maintenance.issues_agent`;
module invocation from the repo root is required for the package imports
(contrib/ itself stays a namespace package without __init__.py). Verified by
diffing the emitted matrix JSON before and after: byte-identical.

One behavior fix: fetch errors for a dataset's issues.json were printed to
stdout, corrupting the matrix JSON GitHub Actions captures; they now simply
yield an empty issue set.
