+++ contrib/maintenance/issues.py
@@ -1,0 +1,67 @@
+"""Normalize, checksum and group the issues emitted by a crawler run."""
+
+import hashlib
+import json
+import re
+from typing import Any
+
+# Python object reprs carry a memory address (e.g. "<Element div at 0x7f...>")
+# that changes every run. The address tells us nothing about which issue this is,
+# so strip it before checksumming or grouping.
+HEX_ADDR_RE = re.compile(r"0x[0-9a-fA-F]+")
+
+
+def issues_checksum(issues: list[Any]) -> str:
+    """Return a checksum that identifies a constant set of crawl issues.
+
+    Use this to seed a deterministic branch name: two crawls that surface the
+    same warnings hash to the same value, so a re-run over an unchanged issue
+    set reuses the branch (and dedupes against an existing PR), while a genuinely
+    changed set produces a new branch.
+
+    Only run-invariant content is hashed. Each issue's `timestamp` and `id` are
+    dropped — `id` is itself a hash that folds in the timestamp, so it churns
+    every run — and memory addresses are scrubbed from the remaining fields. The
+    page-content hashes that live in `data` (e.g. the expected/actual values of a
+    "DOM hash changed" warning) are stable and are kept, so a real source change
+    still moves the checksum.
+    """
+    normalized = [
+        {k: issue.get(k) for k in ("level", "module", "message", "entity", "data")}
+        for issue in issues
+    ]
+    # Sort so the order issues happen to appear in the log doesn't affect the hash.
+    normalized.sort(key=lambda i: json.dumps(i, sort_keys=True, ensure_ascii=False))
+    canonical = json.dumps(normalized, sort_keys=True, ensure_ascii=False)
+    canonical = HEX_ADDR_RE.sub("0x*", canonical)
+    return hashlib.sha1(canonical.encode("utf-8")).hexdigest()
+
+
+# For grouping we normalize more aggressively than for checksumming: values
+# that vary per occurrence (numbers, quoted/bracketed payloads) are collapsed
+# so e.g. every "Assertion ... 379 is not >= 400" instance lands in one bucket.
+_DIGITS_RE = re.compile(r"\d+")
+_QUOTED_RE = re.compile(r"'[^']*'|\"[^\"]*\"")
+
+
+def normalize_message(message: str) -> str:
+    """Collapse an issue message to its run-invariant pattern for grouping."""
+    message = HEX_ADDR_RE.sub("0x*", message)
+    message = _QUOTED_RE.sub("'…'", message)
+    message = _DIGITS_RE.sub("N", message)
+    return message
+
+
+def group_issues(
+    issues: list[dict[str, Any]],
+) -> list[tuple[str, list[dict[str, Any]]]]:
+    """Group issues by normalized message pattern, largest group first.
+
+    Returns (pattern, members) pairs; each members list keeps the original
+    issue dicts so callers can show a representative example.
+    """
+    groups: dict[str, list[dict[str, Any]]] = {}
+    for issue in issues:
+        pattern = normalize_message(str(issue.get("message", "")))
+        groups.setdefault(pattern, []).append(issue)
+    return sorted(groups.items(), key=lambda kv: len(kv[1]), reverse=True)
